added some comments and moved somethings around
