+++ src/location.cpp
@@ -10,6 +10,8 @@
 #include "location.hh"
 #include <iostream>
 
+
+
 // Constructor
 Location::Location() {
     postalCode = "";
